requirements/dev.txt
-r base.txt

pylint-django
autopep8
django-debug-toolbar
coverage